Break out `formatQuery` SQL tests
